Refactor useCanContextualToolbarShow for simplicity and clarity
- Rename to useCanBlockToolbarBeFocused: Each usage of useCanContextualTool
barShow ended up as one blockToolbarCanBeFocused const.
- The scenarios of when the block toolbar can be focused was much simpler t
han the implemented checks. Refactored for the scenarios in which it can be
 focused.
